Jobs: different serialization flavor for AOF.
Now we have two kind of serializations, the usual one used to exchange
messages about jobs with relative expire field, and one with the
absolute expire time (and non overridden state when deserializing) that
is going to be used for AOF and in general when there is to store a job
to retrieve later, so could be used also to send jobs into a local log
when we are out of memory, to consume them later, if this feature is
ever implemented.

In the future the two serialization types may diverge more.