make sure that a randomKey is returned during a GC
A new disque instance without any jobs in memory is running.
If we set the max-memory limit using CONFIG SET max-memory to a
really low limit causes a gc cycle. The disque instance tries now
to free memory using `freeMemoryIfNeeded`. Random keys that can
be freed are returned from the dictionary. If the server has no
jobs there wont be a random key and NULL is returned. This commit
introduces a check to make sure, that a random key is found.