Two different "all nodes reached" code paths unified in tryJobGC().
This commit uses a single code path to check for two cases where all the
nodes are already reached by the ACK and we can GC it ASAP without
sending SETACK commands. Moreover the conditional to check for the
cluster size was modified to use:

    server.cluster->size

Which is more reliable than counting the number of elements in the
server.cluster->nodes dictionary, since the latter will include nodes in
handshake state not actually participating in the cluster.