jscanCallback(): check for dummy job in a consistent way.
So far we always test if a job is a dummy job created by the ACK command
when a job ID is not known by testing if the number of nodes in
nodes_delivered is zero. Currently it is also true that job->queue is
set to NULL in such a case, but it is better to rely on a single
specific thing to test.

Note that testing for zero nodes in the delivery list is also
semantically more sounding, since the concept that the job was never
delivered to any node, from the point of view of the node, makes it
more clearly an unknown job compared to testing a pointer for NULL.