Don't create a dummy ack if the cluster size is 1.
The point of having dummy ACKs is to preserve the information that the
client acknowledged a job even if we are not aware of it: maybe there is
a network partition in progress, so the client was forced to acknowledge
the job to a different node. We'll use this information later to avoid
an additional delivery if the partition heals.

However when the cluster size is 1, we are sure that if the only
existing node does not know about the job, nobody knows. We can avoid
creating the dummy ACK in this case.