Evict delivered jobs having retry set to 0.
Jobs having a retry set to 0 have still to be acknowledged. This way
they remain in the server memory and can be inspected in case of bugs.
However they may also be leaked, because the job can be lost.

In order to make this mechanism working:

1. Jobs with retry = 0 which were already delivered are evicted on
   memory pressure.
2. The Job ID encodes if a job has retry set to 0, so that dummy ACKs
   are not created for those jobs.

For more information please check issue #1 on Github.

Close #1.