Don't serialize JOB_FLAG_DELIVERED in YOURJOBS messages.
JOB_FLAG_DELIVERED is a local flag, it means *this* local node delivered
this message at least once, so must not be propagated. The flag is set
also when YOURJOBS messages are assembled, while fetching retry=0
messages, so if we send messages with this flag to another node, it will
refuse to put the message on the queue again.