Final change to Disque IDs.
Last character was '$'. Given that this character is not selected by
normal GUIs with double click, it was unconfortable to select a Disque
ID with the mouse during development.

Moreover in a fixed len ID, the terminating character is not an important
feature, while the ability to have more reserved bits may turn to be
important. So '$' is now another byte reserved for the future. The final
two byes are now "AA" instead of "A$", however implementations should
never check the last two chars since they may change.