Change Disque versioning scheme.
Disque will use a different versioning scheme compared to Redis,
composed of just major and patchlevel, so versions will be like:

1.0
1.1 (first patch to 1.0)
1.2 (second patch to 1.0)

And so forth. When a new version will be released it will be called 2.0.

Release candidate versions will have a prefix composed of "-rc<number>" So
the first RCs of 1.0 wil be called:

1.0-rc1
1.0-rc2
1.0-rc3