Change REPLJOB / GOTJOB semantics to better handle failures.
After private chat with Sun He (@sunheehnus), talking about ADDJOB command
semantics when synchronous replication is requested, we agreed to change
the way REPLJOB / GOTJOB work in the following ways:

1. REPLJOB is now resend to all the previously contacted nodes, every 50
milliseconds. However if the node already replied with GOTJOB the
REPLJOB is flagged with NOREPLY, since we don't need to waste bandwidth
with a useless reply. Moreover, if there are no additional nodes
involved, the REPLJOB is not resend to nodes that already replied with
GOTJOB.

2. Now nodes receiving REPLJOB always reply with GOTJOB (unless
the message is flagged with NOREPLY), even if they already have a copy
of the message.

What this means is, practically, that there are more chances, during
failures, that ADDJOB is able to successfully replicate the job the
specified number of times, since we'll keep sending REPLJOB and the
receiver will keep replying with GOTJOB multiple times as long as the
specified number of confirmed replicas is not reached.