Fix external replication when REPLICATE is set to 1.
The code checked the wrong conditions to understand if to block for
synchronous replication or not. When replicate is 1, but external
replication is enabled because out of memory, the copy is delivered to
an external node, so we need to block and allocate the data structures
needed to handle synchronous replication.

Thanks to @mathieulongtin for reporting the bug.

Close #148.